backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=12.0,<14.0
python-dotenv==1.0.0
pydantic>=2.11.7,<3.0
pydantic-settings==2.1.0
supabase==2.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pillow==11.0.0
opencv-python-headless==4.9.0.80
google-generativeai==0.3.2
httpx>=0.26,<0.29
email-validator==2.1.0
phonenumbers==8.13.27
redis==5.0.1
celery==5.3.6
boto3==1.34.34

# Payment Integration - DodoPay
requests==2.31.0

# PDF Generation for Certificates
reportlab==4.0.9
pypdf2==3.0.1
weasyprint==60.2

# QR Code for Certificates
qrcode[pil]==7.4.2

# Additional utilities
uuid==1.30